Updated configuration setup
A problem arose with the current cascade of CB's configuration.
Previously, all package variable files were included by default which,
in hindsight, is an oversight. If two or more packages changed the
variables of the same component there would be cascading conflicts with
incorrect values being taken. The new configuration modularizes the
cascade by only including a package's variables if the package is used
in the project. The variable cascade can be controlled on the project
level by the order of the package references. All variables are resolved
before the initialization mixin is called.

Additionally, this patch solved the problem of multiple packages
importing the same shared files.